Fix issue inhibiting multiple multi-user runs.
At the beginning of a run, remove any "end_multi_user.log" files that
may have been created from a previous run. If present, this file will
signal the multi-user run has been performed and not execute it. Thus
it will appear as if the multi-user run was performed when in reality
the previous run's results will be used.